Moved some tests to clusterman and refactored
