+++ cloudman/clusterman/tests/mock_kubectl.py
